Convert nextthink to frame numbers.

+++ src/baseq2/g_local.h
@@ -1024,7 +1024,7 @@ struct edict_s {
     float       yaw_speed;
     float       ideal_yaw;
 
-    float       nextthink;
+    int         nextthink;
     void        (*prethink)(edict_t *ent);
     void        (*think)(edict_t *self);
     void        (*blocked)(edict_t *self, edict_t *other);         // move to moveinfo?
